shoe-store/src/app/dashboard/dashboard.component.ts
import { Component, OnInit } from '@angular/core';
import { Product } from '../product';
import { ProductService } from '../product.service';

@Component({
  selector: 'app-dashboard',
  templateUrl: './dashboard.component.html',
  styleUrls: [ './dashboard.component.css' ]
})
export class DashboardComponent implements OnInit {
  heroes: Product[] = [];

  constructor(private heroService: ProductService) { }

  ngOnInit() {
    this.getHeroes();
  }

  getHeroes(): void {
    this.heroService.getProducts()
      .subscribe(heroes => this.heroes = heroes.slice(1, 5));
  }
}
